Move and modernize operand enums to own header file.
This becomes necessary because the memory operand type
must be ORed with other operand types
(in auto-sync an operand can be a memory and register/immediate
opernad).

So we need to give it a special bit position.
To prevent incorrect assignments it is necessary to make
them assignable to arch operand types.